requirements.txt
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
transformers==4.35.0
torch==2.1.0
plotly==5.17.0
streamlit==1.28.0
pyyaml==6.0.1
wordcloud==1.9.2
pytest==7.4.3
